Feat : LearningMain 화면 API 연결

+++ android/baeuja/app/src/main/java/com/eello/baeuja/ui/screen/learning/LearningCategorySection.kt
@@ -23,7 +23,7 @@ fun LearningCategorySection(
     itemContainerLayoutType: LayoutType,
     classification: ContentClassification,
     learningItems: List<LearningItem>,
-    onNavigateToDetail: (Int) -> Unit = {},
+    onNavigateToDetail: (Long) -> Unit = {},
     onMoreClick: (ContentClassification) -> Unit = {},
     isPreview: Boolean = false
 ) {
