+++ android/baeuja/app/src/main/java/com/eello/baeuja/ui/screen/learning/LearningItemContainer.kt
@@ -11,7 +11,7 @@ fun LearningItemContainer(
     layoutType: LayoutType,
     modifier: Modifier = Modifier,
     items: List<LearningItem>,
-    onNavigateToDetail: (Int) -> Unit = {},
+    onNavigateToDetail: (Long) -> Unit = {},
     isPreview: Boolean = false
 ) {
     when (layoutType) {
